assignment3/pyproject.toml
[build-system]
requires = [
    "setuptools>=61",
    # 4110 only:
    "cython>=3",
    "numpy==1.21.*",
]
build-backend = "setuptools.build_meta"

[project]
version = "1.0.0"
requires-python = ">=3.8"
license = {text = "MIT License"}
name = "in3110_instapy"
description = "This package serves the purpose of changing the color of an image to black and white or sepia"
readme = "README.md"
dependencies = [
    "numpy",
    "numba",
    "pillow",
    "line-profiler",
    "cython",
]


[project.scripts]
in3110_instapy = "instapy.cli:main"

